Revert "updating source code and database"
This reverts commit 17e88a53dd85129c85a787a7918da7ca7487c611.

+++ CodeSource/KitBoxSourceCode/KitBoxSourceCode/Cabinet.cs
@@ -9,9 +9,9 @@ namespace KitBoxSourceCode
         private List<CabinetFloor> cabinetFloors;
         private int cabinetHeight;
 
-        private double cabinetPrice = 0;
+        private int cabinetPrice = 0;
 
-        public double GetCabinetPrice => cabinetPrice;
+        public int GetCabinetPrice => cabinetPrice;
         public int GetCabinetHeight => cabinetHeight;
 
         private Angle angles;
