+++ CodeSource/KitBoxSourceCode/KitBoxSourceCode/IStorageBox.cs
@@ -4,7 +4,7 @@ namespace KitBoxSourceCode
     public interface IStorageBox
     {
         string GetDetails();
-        double GetPrice();
+        int GetPrice();
         int GetHeight();
     }
 }
